requirements.txt
can
can-isotp
cantools
influx_line_protocol
paho-mqtt
